+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+### Added
+- Adding `update-yeoman-generator` integration
+- Removing repository prefix
 
 ## [2.1.0][] - 2017-12-25
 ### Added